Fix and rename range function
The function was called "range", but I was reusing this funciton in
another project and realized it was poorly named. The function does not
generate an arithmetic sequence, at least not directly. It generates a
sequence of numbers within a range which are divisors with what was
previously called "step" now renamed to "divisor".

I also ran prettier on the helper.ts file.